+++ openpnm/contrib/__init__.py
@@ -9,5 +9,5 @@ removed.
 
 """
 
-from _mixed_ip import *
-from _mixed_ip_coop import *
+from ._mixed_ip import *
+from ._mixed_ip_coop import *
