+++ tests/unit/metrics/PorosimetryTest.py
@@ -20,7 +20,7 @@ class PorosimetryTest:
                             model=mod)
 
     def test_no_late_filling(self):
-        mip = op.algorithms.Porosimetry(network=self.net, phase=self.hg)
+        mip = op.metrics.Porosimetry(network=self.net, phase=self.hg)
         mip.set_inlets(pores=self.net.pores('left'))
         mip.run()
         assert len(np.unique(mip['pore.invasion_pressure'])) > 1
@@ -29,7 +29,7 @@ class PorosimetryTest:
         assert len(np.unique(mip['throat.invasion_sequence'])) > 1
 
     def test_late_pore_and_throat_filling(self):
-        mip = op.algorithms.Porosimetry(network=self.net, phase=self.hg)
+        mip = op.metrics.Porosimetry(network=self.net, phase=self.hg)
         mip.set_inlets(pores=self.net.pores('left'))
         # Run without late pore filling
         mip.run()
